etl/requirements.txt
# ETL Dependencies
pandas>=2.0.0
duckdb>=0.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
